requirements.txt
arrow==0.15.5
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-admin-tools==0.9.0
django-background-tasks==1.2.5
django-ckeditor==5.9.0
django-compat==1.0.15
django-js-asset==1.2.2
django-lint==2.0.4
djangorestframework==3.11.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
logilab-astng==0.24.3
logilab-common==1.5.2
mccabe==0.6.1
Pillow==7.0.0
pycodestyle==2.5.0
pylint==0.28.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
